Moved DateRange styled-components to DateRange/styles.js
