--- src/components/styled-components/__tests__/DateRangeComponentContainers.test.js
@@ -1,7 +1,0 @@
-import React from "react";
-import { shallow } from "enzyme";
-import { DateRangeSliderContainer } from "../DateRangeComponentContainer";
-
-test("render DateRangeSliderContainer", () => {
-  shallow(<DateRangeSliderContainer />);
-});
